JIRA-37: added 2 layout variants and conditional rendering for small layout changes
