+++ frontend/MicroBiteClient/package.json
@@ -17,6 +17,7 @@
     "papaparse": "^5.5.2",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-responsive": "^10.0.1",
     "react-router-dom": "^7.4.0",
     "tailwindcss": "^4.0.14"
   },
