+++ frontend/MicroBiteClient/src/header/hooks/useHeaderHook.ts
@@ -1,0 +1,37 @@
+import { useEffect, useRef, useState } from "react";
+import { useAuthContext } from "../../auth/hooks/useAuthContext";
+
+export function useHeaderHook() {
+  const authContext = useAuthContext();
+  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
+  const dropdownRef = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (dropdownRef.current && !dropdownRef.current.contains(event.target as Node)) {
+        setIsDropdownOpen(false);
+      }
+    };
+
+    const handleEscape = (event: KeyboardEvent) => {
+      if (event.key === "Escape") {
+        setIsDropdownOpen(false);
+      }
+    };
+
+    document.addEventListener("mousedown", handleClickOutside);
+    document.addEventListener("keydown", handleEscape);
+
+    return () => {
+      document.removeEventListener("mousedown", handleClickOutside);
+      document.removeEventListener("keydown", handleEscape);
+    };
+  }, []);
+
+  return {
+    authContext,
+    isDropdownOpen,
+    setIsDropdownOpen,
+    dropdownRef,
+  };
+}
